Extract lane panel
